# packages/contract/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "BRS Content API",
    "version": "1.0.0",
    "description": "The BRS content contract. Frontend-agnostic DTOs. Additive-only within a version; breaking changes ship as /v2 with /v1 retained for >= 6 months. MemberDTO deliberately has no contact field (implementation_plan.md §12.1)."
  },
  "servers": [
    {
      "url": "https://api.example.invalid",
      "description": "Set at deploy time"
    }
  ],
  "paths": {
...
    "/v1/stats": {
      "get": {
        "summary": "Computed statistics — the only source of figures on the site",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatsDTO"
                }
              }
            }
          },
          "503": {
            "description": "Insufficient evidence to compute statistics truthfully. Not an outage — returns to 200 once events or achievements exist."
          }
        }
      }
    },
...
    "/v1/collections/{slug}": {
      "get": {
        "summary": "Collection by slug",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CollectionDTO"
                }
              }
            }
          },
          "404": {
            "description": "No such collection"
          }
        }
      }
    },
...
    "/v1/committees/{ordinal}": {
      "get": {
        "summary": "Committee by ordinal",
        "parameters": [
          {
            "name": "ordinal",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommitteeDTO"
                }
              }
            }
          },
          "400": {
            "description": "Ordinal is not an integer"
          },
          "404": {
            "description": "No such committee"
          }
        }
      }
    },
    "/v1/posts/{slug}": {
      "get": {
        "summary": "Post by slug",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PostDTO"
                }
              }
            }
          },
          "404": {
            "description": "No such post"
          }
        }
      }
    },
    "/v1/assets/{id}": {
      "get": {
        "summary": "Redirect to the stored bytes. The stable, provider-agnostic entry point for an asset.",
        "description": "Answers 302 to the object's public URL rather than proxying it, so the façade never sits in the path of image traffic. Storage stays swappable because the redirect target is built from STORAGE_PUBLIC_BASE_URL, not hardcoded. Add ?describe to receive the ImageDTO instead of a redirect.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "describe",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Present at any value: return the ImageDTO instead of redirecting."
          }
        ],
        "responses": {
          "200": {
            "description": "ImageDTO (only when ?describe is present)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImageDTO"
                }
              }
            }
          },
          "302": {
            "description": "Redirect to object storage"
          },
          "400": {
            "description": "Id is not a uuid"
          },
          "404": {
            "description": "No such published asset"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "CollectionDTO": {
        "$schema": "https://json-schema.org/draft/2020-12/schema",
        "type": "object",
        "properties": {
          "slug": {
            "type": "string"
          },
          "label": {
            "type": "string"
          },
          "note": {
            "type": "string"
          },
          "items": {
            "default": [],
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "key": {
                  "type": "string"
                },
                "image": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "url": {
                      "type": "string"
                    },
                    "alt": {
                      "type": "string",
                      "minLength": 12,
                      "maxLength": 280
                    },
                    "width": {
                      "type": "integer",
                      "exclusiveMinimum": 0,
                      "maximum": 9007199254740991
                    },
                    "height": {
                      "type": "integer",
                      "exclusiveMinimum": 0,
                      "maximum": 9007199254740991
                    },
                    "lqip": {
                      "type": "string"
                    },
                    "plate": {
                      "type": "integer",
                      "exclusiveMinimum": 0,
                      "maximum": 9007199254740991
                    },
                    "ratio": {
                      "anyOf": [
                        {
                          "type": "string",
                          "enum": [
                            "1:1",
                            "4:3",
                            "3:2",
                            "16:9",
                            "4:5"
                          ]
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "credit": {
                      "type": "string"
                    },
                    "sources": {
                      "default": [],
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "format": {
                            "type": "string",
                            "enum": [
                              "avif",
                              "webp"
                            ]
                          },
                          "width": {
                            "type": "integer",
                            "exclusiveMinimum": 0,
                            "maximum": 9007199254740991
                          },
                          "url": {
                            "type": "string"
                          }
                        },
                        "required": [
                          "format",
                          "width",
                          "url"
                        ],
                        "additionalProperties": false
                      }
                    }
                  },
                  "required": [
                    "id",
                    "url",
                    "alt",
                    "width",
                    "height",
                    "lqip",
                    "ratio",
                    "sources"
                  ],
                  "additionalProperties": false
                },
                "caption": {
                  "default": [],
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                },
                "title": {
                  "type": "string"
                },
                "year": {
                  "type": "integer",
                  "minimum": -9007199254740991,
                  "maximum": 9007199254740991
                },
                "note": {
                  "type": "string"
                }
              },
              "required": [
                "key",
                "image",
                "caption"
              ],
              "additionalProperties": false
            }
          }
        },
        "required": [
          "slug",
          "label",
          "items"
        ],
        "additionalProperties": false
      },
...
      "CommitteeDTO": {
        "$schema": "https://json-schema.org/draft/2020-12/schema",
        "type": "object",
        "properties": {
          "ordinal": {
            "type": "integer",
            "exclusiveMinimum": 0,
            "maximum": 9007199254740991
          },
          "label": {
            "type": "string"
          },
          "termStart": {
            "type": "integer",
            "minimum": -9007199254740991,
            "maximum": 9007199254740991
          },
          "termEnd": {
            "type": "integer",
            "minimum": -9007199254740991,
            "maximum": 9007199254740991
          },
          "moderator": {
            "type": "object",
            "properties": {
              "name": {
                "type": "string"
              },
              "title": {
                "type": "string"
              },
              "department": {
                "type": "string"
              }
            },
            "required": [
              "name",
              "title",
              "department"
            ],
            "additionalProperties": false
          },
          "groups": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string"
                },
                "members": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "string"
                      },
                      "name": {
                        "type": "string"
                      },
                      "designation": {
                        "type": "string"
                      },
                      "department": {
                        "type": "string"
                      },
                      "batch": {
                        "type": "string"
                      },
                      "committeeOrdinal": {
                        "type": "integer",
                        "exclusiveMinimum": 0,
                        "maximum": 9007199254740991
                      },
                      "team": {
                        "type": "string"
                      },
                      "portrait": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string"
                          },
                          "url": {
                            "type": "string"
                          },
                          "alt": {
                            "type": "string",
                            "minLength": 12,
                            "maxLength": 280
                          },
                          "width": {
                            "type": "integer",
                            "exclusiveMinimum": 0,
                            "maximum": 9007199254740991
                          },
                          "height": {
                            "type": "integer",
                            "exclusiveMinimum": 0,
                            "maximum": 9007199254740991
                          },
                          "lqip": {
                            "type": "string"
                          },
                          "plate": {
                            "type": "integer",
                            "exclusiveMinimum": 0,
                            "maximum": 9007199254740991
                          },
                          "ratio": {
                            "anyOf": [
                              {
                                "type": "string",
                                "enum": [
                                  "1:1",
                                  "4:3",
                                  "3:2",
                                  "16:9",
                                  "4:5"
                                ]
                              },
                              {
                                "type": "null"
                              }
                            ]
                          },
                          "credit": {
                            "type": "string"
                          },
                          "sources": {
                            "default": [],
                            "type": "array",
                            "items": {
                              "type": "object",
                              "properties": {
                                "format": {
                                  "type": "string",
                                  "enum": [
                                    "avif",
                                    "webp"
                                  ]
                                },
                                "width": {
                                  "type": "integer",
                                  "exclusiveMinimum": 0,
                                  "maximum": 9007199254740991
                                },
                                "url": {
                                  "type": "string"
                                }
                              },
                              "required": [
                                "format",
                                "width",
                                "url"
                              ],
                              "additionalProperties": false
                            }
                          }
                        },
                        "required": [
                          "id",
                          "url",
                          "alt",
                          "width",
                          "height",
                          "lqip",
                          "ratio",
                          "sources"
                        ],
                        "additionalProperties": false
                      }
                    },
                    "required": [
                      "id",
                      "name",
                      "designation",
                      "department",
                      "batch",
                      "committeeOrdinal"
                    ],
                    "additionalProperties": false
                  }
                }
              },
              "required": [
                "name",
                "members"
              ],
              "additionalProperties": false
            }
          },
          "isCurrent": {
            "default": false,
            "type": "boolean"
          }
        },
        "required": [
          "ordinal",
          "label",
          "termStart",
          "termEnd",
          "moderator",
          "groups",
          "isCurrent"
        ],
        "additionalProperties": false
      },
...
      "PostDTO": {
        "$schema": "https://json-schema.org/draft/2020-12/schema",
        "type": "object",
        "properties": {
          "slug": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "excerpt": {
            "type": "string"
          },
          "body": {
            "type": "object",
            "properties": {
              "format": {
                "type": "string",
                "enum": [
                  "html",
                  "md"
                ]
              },
              "content": {
                "type": "string"
              }
            },
            "required": [
              "format",
              "content"
            ],
            "additionalProperties": false
          },
          "author": {
            "type": "object",
            "properties": {
              "name": {
                "type": "string"
              },
              "batch": {
                "type": "string"
              }
            },
            "required": [
              "name"
            ],
            "additionalProperties": false
          },
          "publishedAt": {
            "type": "string",
            "description": "ISO-8601 date, e.g. 2024-02-01"
          },
          "tags": {
            "default": [],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "cover": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string"
              },
              "url": {
                "type": "string"
              },
              "alt": {
                "type": "string",
                "minLength": 12,
                "maxLength": 280
              },
              "width": {
                "type": "integer",
                "exclusiveMinimum": 0,
                "maximum": 9007199254740991
              },
              "height": {
                "type": "integer",
                "exclusiveMinimum": 0,
                "maximum": 9007199254740991
              },
              "lqip": {
                "type": "string"
              },
              "plate": {
                "type": "integer",
                "exclusiveMinimum": 0,
                "maximum": 9007199254740991
              },
              "ratio": {
                "anyOf": [
                  {
                    "type": "string",
                    "enum": [
                      "1:1",
                      "4:3",
                      "3:2",
                      "16:9",
                      "4:5"
                    ]
                  },
                  {
                    "type": "null"
                  }
                ]
              },
              "credit": {
                "type": "string"
              },
              "sources": {
                "default": [],
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    "format": {
                      "type": "string",
                      "enum": [
                        "avif",
                        "webp"
                      ]
                    },
                    "width": {
                      "type": "integer",
                      "exclusiveMinimum": 0,
                      "maximum": 9007199254740991
                    },
                    "url": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "format",
                    "width",
                    "url"
                  ],
                  "additionalProperties": false
                }
              }
            },
            "required": [
              "id",
              "url",
              "alt",
              "width",
              "height",
              "lqip",
              "ratio",
              "sources"
            ],
            "additionalProperties": false
          },
          "readingMinutes": {
            "type": "integer",
            "exclusiveMinimum": 0,
            "maximum": 9007199254740991
          }
        },
        "required": [
          "slug",
          "title",
          "excerpt",
          "body",
          "author",
          "publishedAt",
          "tags",
          "readingMinutes"
        ],
        "additionalProperties": false
      },
...
      "StatsDTO": {
        "$schema": "https://json-schema.org/draft/2020-12/schema",
        "type": "object",
        "properties": {
          "committeesDocumented": {
            "type": "integer",
            "minimum": -9007199254740991,
            "maximum": 9007199254740991
          },
          "currentCommitteeSize": {
            "type": "integer",
            "minimum": -9007199254740991,
            "maximum": 9007199254740991
          },
          "workshops": {
            "type": "integer",
            "minimum": -9007199254740991,
            "maximum": 9007199254740991
          },
          "seminars": {
            "type": "integer",
            "minimum": -9007199254740991,
            "maximum": 9007199254740991
          },
          "internationalProgrammes": {
            "type": "integer",
            "minimum": -9007199254740991,
            "maximum": 9007199254740991
          },
          "nationalContests": {
            "type": "integer",
            "minimum": -9007199254740991,
            "maximum": 9007199254740991
          },
          "hostedEventEditions": {
            "type": "integer",
            "minimum": -9007199254740991,
            "maximum": 9007199254740991
          },
          "archivePhotographs": {
            "type": "integer",
            "minimum": -9007199254740991,
            "maximum": 9007199254740991
          },
          "verifiedAwards": {
            "type": "integer",
            "minimum": -9007199254740991,
            "maximum": 9007199254740991
          },
          "earliestEvidenceYear": {
            "type": "integer",
            "minimum": -9007199254740991,
            "maximum": 9007199254740991
          },
          "yearsActive": {
            "type": "integer",
            "minimum": -9007199254740991,
            "maximum": 9007199254740991
          },
          "computedAt": {
            "type": "string",
            "description": "ISO-8601 date, e.g. 2024-02-01"
          }
        },
        "required": [
          "committeesDocumented",
          "currentCommitteeSize",
          "workshops",
          "seminars",
          "internationalProgrammes",
          "nationalContests",
          "hostedEventEditions",
          "archivePhotographs",
          "verifiedAwards",
          "earliestEvidenceYear",
          "yearsActive",
          "computedAt"
        ],
        "additionalProperties": false
      }
    }
  }
}